requirements.txt
torch
torchvision
numpy
pandas
scikit-learn
Pillow
matplotlib
streamlit
optreem